Move service handler to service interface
