+++ conjure-java-core/src/main/java/com/palantir/conjure/java/services/UndertowServiceGenerator.java
@@ -58,8 +58,7 @@ public final class UndertowServiceGenerator implements ServiceGenerator {
             TypeMapper returnTypeMapper) {
         return ImmutableList.of(
                 new UndertowServiceInterfaceGenerator(options)
-                        .generateServiceInterface(serviceDefinition, typeMapper, returnTypeMapper),
-                new UndertowServiceHandlerGenerator(options)
-                        .generateServiceHandler(serviceDefinition, typeDefinitions, typeMapper, returnTypeMapper));
+                        .generateServiceInterface(serviceDefinition, typeDefinitions, typeMapper, returnTypeMapper),
+                new UndertowServiceHandlerGenerator(options).generateServiceHandler(serviceDefinition));
     }
 }
